Fixed issues and unified some things.
Changes:
1)  Fixed:
- fixed issue with redirect to auth.php
- fixed issue with writing data in DB (corrected convert function)
2) Unified
- unified names of classes (all class should start with upper case)
- unified way loading class file (used __autoload instead of include
function)
