+++ acp/loadplugin.php
@@ -13,12 +13,16 @@ if ($target == 'plugin' && preg_match("%^js|css$%", $type))
     $licenseManager = new LicenseManager;
     $allowedPlugins = $licenseManager->checkPermission('allowedPlugins');
     if ($type == 'js')
+    {
         header("Content-type: application/javascript");
+    }
     else
+    {
         header("Content-type: text/css");
+    }
+    
     foreach (getPlugins() as $name)
     {
-
         if (in_array($name, $allowedPlugins))
         {
             $path = "../plugins/$name/acp/$name.$type";
