+++ include/ACL.class.php
@@ -5,11 +5,11 @@ class ACL
   public function get()
   {
     global $sessionId, $response;
-    $masterTable= new table('master');
+    $masterTable= new Table('master');
     $master=$masterTable->loadById( $sessionId );
     $userGroupId=$master['group'];
 
-    $groupTable= new table('group');
+    $groupTable= new Table('group');
     $group=$groupTable->loadById( $userGroupId );
     $userAcl=json_decode($group['acl'], true);
     $response['acl'] = $userAcl;
